backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "clearnote-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "starlette>=0.37.2,<0.38.0",
  "uvicorn[standard]>=0.30.0",
  "gunicorn>=22.0.0",
  "pydantic-settings>=2.3.0",
  "sqlalchemy>=2.0.30",
  "psycopg2-binary>=2.9.9",
  "alembic>=1.13.0",
  "celery[redis]>=5.4.0",
  "boto3>=1.34.0",
  "PyJWT>=2.8.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-postgresql>=6.0.0",
  "pytest-mock>=3.14.0",
  "freezegun>=1.5.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
  "--postgresql-host=localhost",
  "--postgresql-port=5432",
  "--postgresql-user=clearnote",
  "--postgresql-password=clearnote",
  "--postgresql-dbname=clearnote_test",
]
filterwarnings = [
  "ignore:Please use `import python_multipart` instead\\.:PendingDeprecationWarning",
]
